requirements.txt
coloredlogs==15.0.1
numpy==1.24.2
pandas==2.2.3
pyyaml==6.0.2
lightgbm==4.1.0
lazypredict==0.2.12
sklearn==1.5.2
imblearn==0.13.0
